refactor: extract context
