+++ src/components/Chart/Chart.tsx
@@ -10,23 +10,13 @@ import * as React from 'react';
 import { cn } from '@douglasneuroinformatics/libui/utils';
 import * as RechartsPrimitive from 'recharts';
 
-import { ChartConfig } from './types';
+import { ChartContext } from '@/context/ChartContext';
+import { useChart } from '@/hooks/useChart';
+
+import type { ChartConfig } from './types';
+
 // Format: { THEME_NAME: CSS_SELECTOR }
 const THEMES = { dark: '.dark', light: '' } as const;
-
-type ChartContextProps = {
-  config: ChartConfig;
-};
-
-const ChartContext = React.createContext<ChartContextProps | null>(null);
-
-function useChart() {
-  const context = React.useContext(ChartContext);
-  if (!context) {
-    throw new Error('useChart must be used within a <ChartContainer />');
-  }
-  return context;
-}
 
 const ChartContainer = React.forwardRef<
   HTMLDivElement,
