+++ src/hooks/useChart/useChart.ts
@@ -1,0 +1,11 @@
+import { useContext } from 'react';
+
+import { ChartContext } from '@/context/ChartContext';
+
+export function useChart() {
+  const context = useContext(ChartContext);
+  if (!context) {
+    throw new Error('useChart must be used within a <ChartContainer />');
+  }
+  return context;
+}
